Rename fixtures as dummy as they aren't fixtures
