+++ tests/dummy/dummy_class.py
