+++ tests/dummy/dummy_feat.py
@@ -1,0 +1,10 @@
+from charsheets.constants import Feat
+from charsheets.feats.base_feat import BaseFeat
+
+
+#############################################################################
+class DummyFeat(BaseFeat):
+    tag = Feat.NONE
+
+
+# EOF
