+++ tests/dummy/dummy_origin.py
@@ -1,5 +1,6 @@
 from charsheets.origins.base_origin import BaseOrigin
 from charsheets.constants import Feat, Origin, Skill, Stat, Tools
+from tests.dummy.dummy_feat import DummyFeat
 
 
 ###############################################################################
@@ -7,7 +8,7 @@ class DummyOrigin(BaseOrigin):
     __test__ = False
     tag = Origin.NONE
     proficiencies = {Skill.ATHLETICS, Skill.PERCEPTION}
-    origin_feat = Feat.ALERT
+    origin_feat = DummyFeat
     tool_proficiency = Tools.NONE
     origin_stats = (Stat.STRENGTH, Stat.INTELLIGENCE, Stat.WISDOM)
