+++ tests/test_fighter.py
@@ -4,7 +4,7 @@ import unittest
 from charsheets.constants import Skill, Origin, Stat, Ability, Proficiencies
 from charsheets.classes import Fighter, EldritchKnight, Champion, PsiWarrior, BattleMaster, BattleManeuver
 from charsheets.spells import Spells
-from tests.fixtures import DummySpecies, DummyOrigin
+from tests.dummy import DummySpecies, DummyOrigin
 from charsheets.main import render
